Stop supporting nodejs out of the box, but allowing to configure it via a separate module
