+++ korlibs-io-nodejs/module.yaml
@@ -1,0 +1,13 @@
+product:
+  type: lib
+  platforms: [jvm, js, wasm, android, linuxX64, linuxArm64, tvosArm64, tvosX64, tvosSimulatorArm64, macosX64, macosArm64, iosArm64, iosSimulatorArm64, iosX64, watchosArm64, watchosArm32, watchosDeviceArm64, watchosSimulatorArm64, mingwX64]
+  #platforms: [js]
+
+apply: [ ../common.module-template.yaml ]
+
+dependencies:
+  - ../korlibs-io
+
+test-dependencies:
+  - org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.0
+  - ../korlibs-time
