+++ korlibs-util/src@js/korlibs/js/JsExt.kt
@@ -1,5 +1,7 @@
 package korlibs.js
 
+import org.khronos.webgl.*
+
 fun <T> JSIterable<T>.toArray(): Array<T> {
     return Array_from(this)
 }
@@ -51,4 +53,14 @@ external class JsArray<T> {
     companion object {
         fun from(value: dynamic): Array<dynamic>
     }
+}
+
+fun Uint8Array.toByteArray(): ByteArray = Int8Array(this.unsafeCast<Int8Array>()).unsafeCast<ByteArray>()
+fun Int8Array.toByteArray(): ByteArray = this.unsafeCast<ByteArray>()
+//fun ByteArray.toNodeJsBufferU8(): NodeBuffer = Uint8Array(this.unsafeCast<ArrayBuffer>()).asDynamic()
+
+fun ByteArray.asInt8Array(): Int8Array = this.unsafeCast<Int8Array>()
+fun ByteArray.toUint8Array(): Uint8Array {
+    val i = this.asInt8Array()
+    return Uint8Array(i.buffer, i.byteOffset, i.length)
 }
